Moved Table components to Table folder
